Implemented Parceable to Movie class to transfer Movie data in intent. Added intent to Main Activity. Updated MovieDetailsActivity with intent and set Movie data to appropriate views. Added up button functionality. Updated colors.

+++ app/src/main/java/com/example/android/popularmovies/utilities/MovieDBJsonUtils.java
@@ -33,7 +33,7 @@ public class MovieDBJsonUtils {
                 String movieTitle = resultsObject.getString(TITLE_KEY);
                 String moviePosterPath = "http://image.tmdb.org/t/p/w185" + resultsObject.getString(POSTER_PATH_KEY);
                 String movieOverview = resultsObject.getString(OVERVIEW_KEY);
-                long movieVoteAvg = resultsObject.getLong(VOTE_AVG_KEY);
+                Double movieVoteAvg = resultsObject.getDouble(VOTE_AVG_KEY);
                 String movieReleaseDate = resultsObject.getString(RELEASE_DATE_KEY);
 
                 Movie newMovie = new Movie(movieTitle, movieReleaseDate, moviePosterPath, movieVoteAvg, movieOverview);
